refactor DcampBlockBase and add Newsletter block for footer
